requirements.txt
# Python Dependencies for Financial Asset Analysis

# Core Data Processing
pyspark==3.5.0
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Financial Data
yfinance==0.2.32
pandas-datareader==0.10.0
yahoo-fin==0.8.9.1
finnhub-python==2.4.19

# Machine Learning - Supervised
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2

# Deep Learning
tensorflow==2.15.0
keras==2.15.0
torch==2.1.1
transformers==4.35.2

# Time Series Analysis
statsmodels==0.14.0
prophet==1.1.5
pmdarima==2.0.4

# Clustering and Unsupervised
hdbscan==0.8.33
umap-learn==0.5.5

# Dashboard and Visualization
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
dash-table==5.0.0
dash-extensions==1.0.13

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Export and Reporting
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
kaleido==0.2.1
python-pptx==0.6.23
Pillow==10.1.0

# Database and Storage
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3.post1
schedule==1.2.0
tqdm==4.66.1

# Monitoring and Logging
prometheus-client==0.19.0
py4j==0.10.9.7

# Performance
joblib==1.3.2
numba==0.58.1
bottleneck==1.3.7

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
